requirements.txt
django>=2.1.6
mysqlclient==1.3.13
pycodestyle==2.4
django-mptt==0.8.6
Pillow==5.3.0
django-crispy-forms==1.7.2
django-notifications-hq==1.5.0
django-cleanup==3.0.1
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
django-rest-swagger==2.2.0
social-auth-app-django==3.1.0
python-social-auth[django]==0.3.6
django-countries==5.3.2 
django-import-export